Variety of Git repo supports + HUGO_PUSH_SECRET

+++ deploy.sh
@@ -1,7 +1,16 @@
 #!/bin/ash
 
+set -e
+
+refBranch=$(basename $WEBHOOK_REF)
+
+if [ "$HUGO_BRANCH" != "$refBranch" ]; then
+    echo "Wrong branch pushed '$refBranch'. Cancelling build.";
+    return 1;
+fi
+
 # Fetch website
-git clone -b $HUGO_BRANCH $HUGO_REPOSITORY --recurse-submodules /temp/website
+git clone -b $HUGO_BRANCH $WEBHOOK_REPOSITORY --recurse-submodules /temp/website
 
 # Clear old website files
 rm -R /www/*
